backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
cryptography
groq
python-multipart
python-dotenv
psycopg2-binary
